+++ packages/api/schema/types/product/product.js
@@ -136,7 +136,7 @@ export default [
       isNetPrice: Boolean!
       country: Country!
       currency: Currency!
-      amount: Int!
+      amount: BigInt!
       maxQuantity: Int
     }
   `,
